add entry_point for CellMetaData
