+++ src/py/jupyter_graphql/schema/contents/nbformat/cell_metadata.py
@@ -1,0 +1,61 @@
+import graphene as G
+
+from ..entry_points import load_cell_metadata_bases
+
+
+class CellMetaData(*load_cell_metadata_bases(), G.Interface):
+    name = G.String()
+    tags = G.List(G.NonNull(G.String))
+
+
+# Official Jupyter Metadata
+
+
+class HideableSource(G.Interface):
+    source_hidden = G.Boolean()
+
+
+class HideableOutputs(G.Interface):
+    outputs_hidden = G.Boolean()
+
+
+class CodeCellJupyterMetaData(G.ObjectType):
+    class Meta:
+        interfaces = (HideableSource, HideableOutputs)
+
+
+class RawCellJupyterMetaData(G.ObjectType):
+    class Meta:
+        interfaces = (HideableSource,)
+
+
+class MarkdownCellJupyterMetaData(G.ObjectType):
+    class Meta:
+        interfaces = (HideableSource,)
+
+
+# Cell Metadata
+
+
+class CodeCellMetaData(G.ObjectType):
+    class Meta:
+        interfaces = (CellMetaData,)
+
+    collapsed = G.Boolean()
+    scrolled = G.Boolean()
+    jupyter = G.Field(CodeCellJupyterMetaData)
+
+
+class RawCellMetaData(G.ObjectType):
+    class Meta:
+        interfaces = (CellMetaData,)
+
+    format = G.String()
+    jupyter = G.Field(RawCellJupyterMetaData)
+
+
+class MarkdownCellMetaData(G.ObjectType):
+    class Meta:
+        interfaces = (CellMetaData,)
+
+    jupyter = G.Field(MarkdownCellJupyterMetaData)
