+++ src/py/jupyter_graphql/schema/contents/nbformat/cells.py
@@ -1,5 +1,8 @@
 import graphene as G
 
+from .output import Media, Output, OUTPUT_TYPES
+from .cell_metadata import CodeCellMetaData, MarkdownCellMetaData, RawCellMetaData
+
 
 class CellType(G.Enum):
     code = "code"
@@ -14,74 +17,6 @@ class Cell(G.Interface):
     @classmethod
     def resolve_type(cls, it, info):
         return CELL_TYPES[it["cell_type"]]
-
-
-class CellMetaData(G.Interface):
-    name = G.String()
-    tags = G.List(G.NonNull(G.String))
-
-
-# Official Jupyter Metadata
-
-
-class HideableSource(G.Interface):
-    source_hidden = G.Boolean()
-
-
-class HideableOutputs(G.Interface):
-    outputs_hidden = G.Boolean()
-
-
-class CodeCellJupyterMetaData(G.ObjectType):
-    class Meta:
-        interfaces = (HideableSource, HideableOutputs)
-
-
-class RawCellJupyterMetaData(G.ObjectType):
-    class Meta:
-        interfaces = (HideableSource,)
-
-
-class MarkdownCellJupyterMetaData(G.ObjectType):
-    class Meta:
-        interfaces = (HideableSource,)
-
-# Cell Metadata
-
-
-class CodeCellMetaData(G.ObjectType):
-    class Meta:
-        interfaces = (CellMetaData,)
-
-    collapsed = G.Boolean()
-    scrolled = G.Boolean()
-    jupyter = G.Field(CodeCellJupyterMetaData)
-
-
-class RawCellMetaData(G.ObjectType):
-    class Meta:
-        interfaces = (CellMetaData,)
-
-    format = G.String()
-    jupyter = G.Field(RawCellJupyterMetaData)
-
-
-class MarkdownCellMetaData(G.ObjectType):
-    class Meta:
-        interfaces = (CellMetaData,)
-
-    jupyter = G.Field(MarkdownCellJupyterMetaData)
-
-
-class Media(G.ObjectType):
-    mimetype = G.String()
-    data = G.types.json.JSONString()
-
-    def resolve_mimetype(it, info):
-        return it[0]
-
-    def resolve_data(it, info):
-        return it[1]
 
 
 class Attachment(G.ObjectType):
@@ -100,71 +35,6 @@ class HasAttachments(G.Interface):
 
     def resolve_attachments(it, info):
         return it.get("attachments").items()
-
-
-# Output
-
-
-class OutputType(G.Enum):
-    execute_result = "execute_result"
-    display_data = "display_data"
-    stream = "stream"
-    error = "error"
-
-
-class Output(G.Interface):
-    output_type = G.Field(OutputType)
-
-    @classmethod
-    def resolve_type(cls, it, info):
-        return OUTPUT_TYPES[it["output_type"]]
-
-
-class RichOutput(G.Interface):
-    data = G.List(Media)
-    metadata = G.types.json.JSONString()
-
-    def resolve_data(it, info):
-        return it["data"].items()
-
-
-class ExecuteResult(G.ObjectType):
-    class Meta:
-        interfaces = (Output, RichOutput)
-
-    execution_count = G.Int()
-
-
-class DisplayData(G.ObjectType):
-    class Meta:
-        interfaces = (Output, RichOutput)
-
-
-class StreamType(G.Enum):
-    stdout = "stdout"
-    stderr = "stderr"
-
-
-class Stream(G.ObjectType):
-    class Meta:
-        interfaces = (Output,)
-
-    name = G.Field(StreamType)
-    text = G.String()
-
-
-class Error(G.ObjectType):
-    class Meta:
-        interfaces = (Output,)
-
-    ename = G.String()
-    evalue = G.String()
-    traceback = G.List(G.String)
-
-
-OUTPUT_TYPES = dict(
-    execute_result=ExecuteResult, display_data=DisplayData, stream=Stream, error=Error
-)
 
 
 # The actual Cells
@@ -193,6 +63,7 @@ class RawCell(G.ObjectType):
     metadata = G.Field(RawCellMetaData)
 
 
+# used to resolve type
 CELL_TYPES = dict(code=CodeCell, markdown=MarkdownCell, raw=RawCell)
